requirements.txt
biopython==1.73
certifi==2018.11.29
chardet==3.0.4
idna==2.8
lxml==4.3.1
numpy==1.16.1
requests==2.21.0

Unidecode==1.0.23
urllib3==1.24.1
